+++ app/javascript/staff/assets/images/vue.png
