+++ config/webpack/utils/aliases.js
@@ -1,0 +1,12 @@
+const path = require('path');
+
+module.exports = {
+  resolve: {
+    alias: {
+      staff_images: path.resolve(
+        __dirname,
+        '../../../app/javascript/staff/assets/images',
+      ),
+    },
+  },
+};
